+++ src/main/java/com/example/event/service/OrganizerServiceImpl.java
@@ -1,7 +1,10 @@
 package com.example.event.service;
 
 import com.example.event.entity.Organizer;
+import com.example.event.error.EventOrganizerException;
+import com.example.event.error.OrganizerDoesNotExistException;
 import com.example.event.repository.OrganizerRepository;
+import org.aspectj.weaver.ast.Or;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -22,17 +25,41 @@ public class OrganizerServiceImpl implements OrganizerService{
     }
 
     @Override
-    public void deleteOrganizer(Long id) {
+    public void deleteOrganizer(Long id) throws OrganizerDoesNotExistException {
+        if(organizerRepository.findById(id).isEmpty())
+        {
+            throw new OrganizerDoesNotExistException("Organizer does not exist");
+        }
         organizerRepository.deleteById(id);
     }
 
     @Override
-    public Organizer updateOrganizer(Organizer organizer) {
-        return organizerRepository.save(organizer);
+    public Organizer updateOrganizer(Organizer organizer, Long id)
+            throws OrganizerDoesNotExistException {
+
+        if(organizerRepository.findById(id).isEmpty())
+        {
+            throw new OrganizerDoesNotExistException("Organizer does not exist");
+        }
+
+        Organizer previousOrganizer = organizerRepository.findById(id).get();
+
+        previousOrganizer.setName(organizer.getName());
+        previousOrganizer.setEmail(organizer.getEmail());
+        previousOrganizer.setPhone(organizer.getPhone());
+
+
+        return organizerRepository.save(previousOrganizer);
     }
 
     @Override
-    public Organizer getOrganizer(Long id) {
+    public Organizer getOrganizer(Long id) throws EventOrganizerException {
+
+        if(organizerRepository.findById(id).isEmpty())
+        {
+            throw new EventOrganizerException("Organizer does not exist");
+        }
+
         return organizerRepository.findById(id).get();
     }
